Normalize app URLs (localhost<>127.0.0.1) at parse time
This ensures that functions always have a correct deterministic ID
